+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Maintain Changelog
 
-## Master
+## 0.3.0 (2018-09-11)
 
 ### Breaking